Release/0.0.59

+++ lib/pages/home/bottom_sheets/post_actions.dart
@@ -1,13 +1,14 @@
 import 'package:Okuna/models/post.dart';
 import 'package:Okuna/models/user.dart';
+import 'package:Okuna/pages/home/bottom_sheets/rounded_bottom_sheet.dart';
 import 'package:Okuna/provider.dart';
+import 'package:Okuna/services/bottom_sheet.dart';
 import 'package:Okuna/services/localization.dart';
 import 'package:Okuna/services/modal_service.dart';
 import 'package:Okuna/services/toast.dart';
 import 'package:Okuna/services/user.dart';
 import 'package:Okuna/services/httpie.dart';
 import 'package:Okuna/widgets/icon.dart';
-import 'package:Okuna/widgets/theming/primary_color_container.dart';
 import 'package:Okuna/widgets/theming/text.dart';
 import 'package:Okuna/widgets/tiles/actions/close_post_tile.dart';
 import 'package:Okuna/widgets/tiles/actions/disable_comments_post_tile.dart';
@@ -25,16 +26,15 @@ class OBPostActionsBottomSheet extends StatefulWidget {
   final Function onCommunityExcluded;
   final Function onUndoCommunityExcluded;
 
-  const OBPostActionsBottomSheet(
-      {Key key,
-      @required this.post,
-      @required this.onPostReported,
-      @required this.onPostDeleted,
-      this.onCommunityExcluded,
-      this.onUndoCommunityExcluded,
-      this.isTopPost = false,
-      })
-      : super(key: key);
+  const OBPostActionsBottomSheet({
+    Key key,
+    @required this.post,
+    @required this.onPostReported,
+    @required this.onPostDeleted,
+    this.onCommunityExcluded,
+    this.onUndoCommunityExcluded,
+    this.isTopPost = false,
+  }) : super(key: key);
 
   @override
   OBPostActionsBottomSheetState createState() {
@@ -47,6 +47,7 @@ class OBPostActionsBottomSheetState extends State<OBPostActionsBottomSheet> {
   ModalService _modalService;
   ToastService _toastService;
   LocalizationService _localizationService;
+  BottomSheetService _bottomSheetService;
 
   @override
   Widget build(BuildContext context) {
@@ -55,6 +56,7 @@ class OBPostActionsBottomSheetState extends State<OBPostActionsBottomSheet> {
     _modalService = openbookProvider.modalService;
     _toastService = openbookProvider.toastService;
     _localizationService = openbookProvider.localizationService;
+    _bottomSheetService = openbookProvider.bottomSheetService;
 
     User loggedInUser = _userService.getLoggedInUser();
 
@@ -109,7 +111,7 @@ class OBPostActionsBottomSheetState extends State<OBPostActionsBottomSheet> {
             postActions.add(ListTile(
               leading: const OBIcon(OBIcons.editPost),
               title: OBText(
-               _localizationService.post__edit_title,
+                _localizationService.post__edit_title,
               ),
               onTap: _onWantsToEditPost,
             ));
@@ -131,8 +133,7 @@ class OBPostActionsBottomSheetState extends State<OBPostActionsBottomSheet> {
             ));
           }
 
-          return OBPrimaryColorContainer(
-            mainAxisSize: MainAxisSize.min,
+          return OBRoundedBottomSheet(
             child: Column(
               children: postActions,
               mainAxisSize: MainAxisSize.min,
@@ -142,20 +143,22 @@ class OBPostActionsBottomSheetState extends State<OBPostActionsBottomSheet> {
   }
 
   Future _onWantsToDeletePost() async {
-    try {
-      await _userService.deletePost(widget.post);
-      _toastService.success(message: _localizationService.post__actions_deleted, context: context);
-      widget.onPostDeleted(widget.post);
-      Navigator.pop(context);
-    } catch (error) {
-      _onError(error);
-    }
+    _bottomSheetService.showConfirmAction(
+        context: context,
+        subtitle: _localizationService.post__actions_delete_description,
+        actionCompleter: (BuildContext context) async {
+          await _userService.deletePost(widget.post);
+          _toastService.success(
+              message: _localizationService.post__actions_deleted,
+              context: context);
+          widget.onPostDeleted(widget.post);
+        });
   }
 
   Future _onWantsToEditPost() async {
     try {
       await _modalService.openEditPost(context: context, post: widget.post);
-      Navigator.pop(context);
+      _dismiss();
     } catch (error) {
       _onError(error);
     }
@@ -169,13 +172,14 @@ class OBPostActionsBottomSheetState extends State<OBPostActionsBottomSheet> {
       String errorMessage = await error.toHumanReadableMessage();
       _toastService.error(message: errorMessage, context: context);
     } else {
-      _toastService.error(message: _localizationService.error__unknown_error, context: context);
+      _toastService.error(
+          message: _localizationService.error__unknown_error, context: context);
       throw error;
     }
   }
 
   void _dismiss() {
-    Navigator.pop(context);
+    _bottomSheetService.dismissActiveBottomSheet(context: context);
   }
 }
