+++ lib/pages/home/bottom_sheets/post_comment_more_actions.dart
@@ -1,14 +1,15 @@
 import 'package:Okuna/models/post.dart';
 import 'package:Okuna/models/post_comment.dart';
 import 'package:Okuna/models/user.dart';
+import 'package:Okuna/pages/home/bottom_sheets/rounded_bottom_sheet.dart';
 import 'package:Okuna/provider.dart';
+import 'package:Okuna/services/bottom_sheet.dart';
 import 'package:Okuna/services/localization.dart';
 import 'package:Okuna/services/modal_service.dart';
 import 'package:Okuna/services/navigation_service.dart';
 import 'package:Okuna/services/toast.dart';
 import 'package:Okuna/services/user.dart';
 import 'package:Okuna/widgets/icon.dart';
-import 'package:Okuna/widgets/theming/primary_color_container.dart';
 import 'package:Okuna/widgets/theming/text.dart';
 import 'package:Okuna/widgets/tiles/actions/mute_post_comment_tile.dart';
 import 'package:async/async.dart';
@@ -41,6 +42,7 @@ class OBPostCommentMoreActionsBottomSheetState
   NavigationService _navigationService;
   LocalizationService _localizationService;
   ModalService _modalService;
+  BottomSheetService _bottomSheetService;
   bool _requestInProgress;
   CancelableOperation _requestOperation;
 
@@ -64,9 +66,9 @@ class OBPostCommentMoreActionsBottomSheetState
     _navigationService = provider.navigationService;
     _localizationService = provider.localizationService;
     _modalService = provider.modalService;
+    _bottomSheetService = provider.bottomSheetService;
 
-    return OBPrimaryColorContainer(
-      mainAxisSize: MainAxisSize.min,
+    return OBRoundedBottomSheet(
       child: Opacity(
         opacity: _requestInProgress ? 0.5 : 1,
         child: Column(
@@ -86,26 +88,15 @@ class OBPostCommentMoreActionsBottomSheetState
     ];
     User loggedInUser = _userService.getLoggedInUser();
 
-    if (loggedInUser.canDeletePostComment(widget.post, widget.postComment)) {
-      actionTiles.add(
-        ListTile(
-          leading: const OBIcon(OBIcons.deletePost),
-          title: OBText(
-            _localizationService.post__actions_delete_comment,
-          ),
-          onTap: _deletePostComment,
-        ),
-      );
-    }
-
     if (loggedInUser.canReportPostComment(widget.postComment)) {
       actionTiles.add(
         Opacity(
           opacity: widget.postComment.isReported ?? false ? 0.5 : 1,
           child: ListTile(
             leading: const OBIcon(OBIcons.report),
-            title: OBText(
-                widget.postComment.isReported ?? false ? _localizationService.post__actions_reported_text : _localizationService.post__actions_report_text),
+            title: OBText(widget.postComment.isReported ?? false
+                ? _localizationService.post__actions_reported_text
+                : _localizationService.post__actions_report_text),
             onTap: _reportPostComment,
           ),
         ),
@@ -120,6 +111,18 @@ class OBPostCommentMoreActionsBottomSheetState
             _localizationService.post__actions_edit_comment,
           ),
           onTap: _editPostComment,
+        ),
+      );
+    }
+
+    if (loggedInUser.canDeletePostComment(widget.post, widget.postComment)) {
+      actionTiles.add(
+        ListTile(
+          leading: const OBIcon(OBIcons.deletePost),
+          title: OBText(
+            _localizationService.post__actions_delete_comment,
+          ),
+          onTap: _onWantsToDeletePostComment,
         ),
       );
     }
@@ -139,46 +142,30 @@ class OBPostCommentMoreActionsBottomSheetState
         });
   }
 
-  void _deletePostComment() async {
+  Future _onWantsToDeletePostComment() async {
     if (_requestInProgress) return;
     _setRequestInProgress(true);
-    _dismissMoreActions();
-    try {
-      _requestOperation = CancelableOperation.fromFuture(
-          _userService.deletePostComment(
-              postComment: widget.postComment, post: widget.post));
-
-      await _requestOperation.value;
-      if (widget.postComment.parentComment == null)
-        widget.post.decreaseCommentsCount();
-      _toastService.success(message: _localizationService.post__actions_comment_deleted, context: context);
-      if (widget.onPostCommentDeleted != null) {
-        widget.onPostCommentDeleted(widget.postComment);
-      }
-    } catch (error) {
-      _onError(error);
-    } finally {
-      _setRequestInProgress(false);
-    }
+    _bottomSheetService.showConfirmAction(
+        context: context,
+        subtitle: _localizationService.post__actions_delete_comment_description,
+        actionCompleter: (BuildContext context) async {
+          await _userService.deletePostComment(
+              postComment: widget.postComment, post: widget.post);
+          _toastService.success(
+              message: _localizationService.post__actions_comment_deleted,
+              context: context);
+          if (widget.postComment.parentComment == null)
+            widget.post.decreaseCommentsCount();
+          if (widget.onPostCommentDeleted != null) {
+            widget.onPostCommentDeleted(widget.postComment);
+          }
+        });
   }
 
   void _editPostComment() async {
     _dismissMoreActions();
     await _modalService.openExpandedCommenter(
         context: context, post: widget.post, postComment: widget.postComment);
-  }
-
-  void _onError(error) async {
-    if (error is HttpieConnectionRefusedError) {
-      _toastService.error(
-          message: error.toHumanReadableMessage(), context: context);
-    } else if (error is HttpieRequestError) {
-      String errorMessage = await error.toHumanReadableMessage();
-      _toastService.error(message: errorMessage, context: context);
-    } else {
-      _toastService.error(message: _localizationService.error__unknown_error, context: context);
-      throw error;
-    }
   }
 
   void _setRequestInProgress(bool requestInProgress) {
@@ -188,6 +175,6 @@ class OBPostCommentMoreActionsBottomSheetState
   }
 
   void _dismissMoreActions() {
-    Navigator.pop(context);
+    _bottomSheetService.dismissActiveBottomSheet(context: context);
   }
 }
